+++ rapidapi/work_with_city.py
@@ -1,0 +1,17 @@
+from rapidapi.get_info import api_request
+
+
+def city_search(city_name):
+    query_string = {'q': city_name, 'locale': 'en_US'}
+    response = api_request(method_endswith='locations/v3/search',
+                           params=query_string,
+                           method_type='GET')
+    if response:
+        cities = list()
+        for i in response['sr']:
+            if i['type'] == "CITY":
+                cities.append(
+                    dict(id=i['gaiaId'],
+                         region_name=i['regionNames']['fullName'])
+                )
+        return cities
