+++ states/bot_states.py
@@ -15,6 +15,7 @@ class MyStates(StatesGroup):
     check_out = State()
     city_detail = State()
     location_confirmation = State()
+    bestdeal = State()
     bestdeal_print_results = State()
     bestdeal_distance_min_flag = State()
     bestdeal_distance_max_flag = State()
